chore: move logic for machine network config updates to new file
This mirrors what is in the service layer.
